Script/requirements.txt
astroid==1.6.6
autopep8==1.5.4
backports.functools-lru-cache==1.6.1
blurhash==1.1.4
boost==0.1
bson==0.5.10
certifi==2020.6.20
chardet==3.0.4
cocos2d==0.6.8
colorama==0.4.4
configparser==4.0.2
coverage==5.3
Cython==0.28.2
decorator==4.4.2
enum34==1.1.10
first==2.0.2
freetype-py==2.1.0.post1
funcsigs==1.0.2
future==0.18.2
futures==3.3.0
glfw==2.0.0
graphviz==0.14.2
idna==2.10
imgui==1.2.0
inputs==0.5
isort==4.3.21
Jinja2==2.11.2
lazy-object-proxy==1.5.1
MarkupSafe==1.1.1
Mastodon.py==1.5.1
mccabe==0.6.1
memory-profiler==0.57.0
mock==3.0.5
msgpack==1.0.0
numpy==1.16.6
objgraph==3.5.0
OpenGLContext==2.3.0b1
Pillow==6.2.2
protobuf==3.12.2
psutil==5.7.3
pyassimp==4.1.4
pybullet==1.0.0
pycallgraph==1.0.1
pycodestyle==2.6.0
PyDispatcher==2.0.5
pygame==1.9.6
pyglet==1.4.10
PyGLM==0.4.8b1
pylint==1.9.5
PyOpenGL @ file:///D:/_Downloads/PyOpenGL-3.1.5-cp27-cp27m-win32.whl
PyOpenGL-accelerate @ file:///D:/MiniProj_01/dist/PyOpenGL_accelerate-3.1.5-cp27-cp27m-win32.whl
PySDL2==0.9.7
python-dateutil==2.8.1
python-magic==0.4.18
pytz==2020.1
PyVRML97==2.3.1
requests==2.24.0
SimpleParse==2.2.2
singledispatch==3.4.0.3
six==1.15.0
SQLAlchemy==1.3.19
tabulate==0.8.7
toml==0.10.1
typing==3.7.4.3
urllib3==1.25.10
wrapt==1.12.1
